requirements.txt
requests
scrapy
pandas
colorama
datetime
yfinance
psycopg2-binary
python-dotenv
watchdog
flask
redis
psutil
scrapy-cloudflare-middleware